requirements.txt
Flask==2.3.3
Flask-CORS==3.0.10
requests==2.31.0
googletrans==4.0.0-rc1
python-dotenv==1.0.0
